Enable batch replacing for strings
Allows the replace function to be called with two arguments instead of three,
if the second argument is a dictionary.
In this case, the dictionary is used to perform a series of replacements,
replacing each occurrence of each key with the corresponding value. Keys can
be either string literals or regular expressions.

Example:
  >>> replace("asdf", {'a': 'b', 's': 'x', /[df]/: '.'})
  "bx.."

Note that care must be taken to ensure that the value of one mapping does not
match the key of another, or this functionality may behave in unexpected ways.